Add files via upload
This Java application is designed to let the user keep track of who they have interacted with and for how long, input additional information about them, and view the information later.  This lets users know who they have been around if exposed to covid and know who was more likely to contract the virus.